refactor(ui): Simplify TailwindCSS class access
Simplify connecting different statuses to TailwindCSS color classes by
merging the handling from two separate files into one.

As this file exports also other than color classes, rename the file 
accordingly.
