+++ ui/src/helpers/calc-overall-vulnerability.ts
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-import { VulnerabilityRating } from '@/types/status-types-and-constants';
+import { VulnerabilityRating } from '@/helpers/get-status-class';
 
 // Calculate the overall vulnerability rating based on the individual ratings,
 // which are input as numbers 0-10.
