+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/_layout/runs/$runIndex/config/-components/advisor-job-details.tsx
@@ -22,7 +22,7 @@ import { Badge } from '@/components/ui/badge';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Label } from '@/components/ui/label';
 import { calculateDuration } from '@/helpers/get-run-duration';
-import { getStatusBackgroundColor } from '@/helpers/get-status-colors';
+import { getStatusBackgroundColor } from '@/helpers/get-status-class';
 
 type AdvisorJobDetailsProps = {
   run: OrtRun;
