+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/_layout/runs/$runIndex/index.tsx
@@ -42,7 +42,7 @@ import { calculateDuration } from '@/helpers/get-run-duration';
 import {
   getStatusBackgroundColor,
   getStatusFontColor,
-} from '@/helpers/get-status-colors';
+} from '@/helpers/get-status-class';
 import { toast } from '@/lib/toast';
 
 const RunComponent = () => {
